// integrations/grafana/src/gen/grafana-legacy-client/types.gen.ts
export type MatchRegexps = {
  [key: string]: string
}

/**
 * MatchType is an enum for label matching types.
 */
export type MatchType = number

/**
 * Matcher models the matching of a label.
 */
export type Matcher = {
  Name?: string
  Type?: MatchType
  Value?: string
}

/**
 * Matchers is a slice of Matchers that is sortable, implements Stringer, and
 * provides a Matches method to match a LabelSet against all Matchers in the
 * slice. Note that some users of Matchers might require it to be sorted.
 */
export type Matchers = Array<Matcher>
// ...
export type ObjectMatcher = Array<string>

/**
 * ObjectMatchers is a list of matchers that can be used to filter alerts.
 */
export type ObjectMatchers = Array<ObjectMatcher>
// ...
export type Provenance = string
// ...
export type Route = {
  active_time_intervals?: Array<string>
  continue?: boolean
  group_by?: Array<string>
  group_interval?: string
  group_wait?: string
  /**
   * Deprecated. Remove before v1.0 release.
   */
  match?: {
    [key: string]: string
  }
  match_re?: MatchRegexps
  matchers?: Matchers
  mute_time_intervals?: Array<string>
  object_matchers?: ObjectMatchers
  provenance?: Provenance
  receiver?: string
  repeat_interval?: string
  routes?: Array<Route>
}